feat: administrator dashboard, counsellor and module views (#17)
* CU-251f194 - Administrator dashboard side menu, header and login/logout

* CU-251f194 - Administrator dashboard side menu, header and login/logout

* feat: implement ui for counsellor list page

includes the HTML and styles for the list, and TODOs where additional
functionality is to be implemented in future.

* feat: implement sort counsellor list function

counsellor list will be sorted by counsellor names on click of the sort
icon. also blur sort icon on hover to better indicate that it is click-able.

* feat: implement filtering of counsellor list by user-inputted term

as the user inputs the search term, filter the counsellor list by
searching for the term (ignoring case) in counsellor names and
counsellor positions.

* style: format counsellor-list.svelte

run the formatter on the file, as some code lines were too long, and fix
whitespace.

* Made the navbar, user button and side drawer

* placed links in head element in app.html for bootstrap

* chore: install tailwindcss

* chore:: install daisyui dependency

* feat: add daisy-spike page

* chore: run prettier

* adjusted admin bar to use daisy ui

* build: remove sveltestrap from package and package-lock

as per announcement on discord, we are no longer using sveltestrap. the
sveltestrap (and other unused dependencies) were re-added back in when we merged the branch with rc.

* fix: switch to tailwind in counsellor list

also convert counsellor list code into a component

* style: run prettier on AdminBar file and rename file

components should start with capital

* refactor: generalise the counsellor list

instead of a 'CounsellorList' component, we will have a 'FilterableList'
component that can be used across both the 'Counsellors' page and the
'Modules' page.

* feat: implement 'Modules' page

also change the images in the FilterableList to a rounded-rectangle
shape as the default.

* dashboard list of counsellors implementation

first attempt at councellor list on dashboard still needs styling and components extracted

* made counsellor pictures rounded

made pictures rounded

* fixed up style a bit and added text as buttons

* Made counsellor page and modules page to put frontend code together for naviagtion later

* feat: implement dashboard modules

* fix import reference

* changed icon of account button on header

* refactor: clean-up the admin routes and combine code

also use formatter to clean-up code style.

* feat: add navigation links to link the admin pages

link the 'See all' buttons and side menu items to their corresponding
pages.

* feat: refine styling of the admin pages

* Index had dashboard gone so i placed it back in

* removed tiny dash in the index.html

* Made changes according to pull request

* fixed import issues in the files

* moved components out of admin folder except AdminBar

* chore: run prettier

Co-authored-by: Jason Tay <jtay0034@student.monash.edu>
Co-authored-by: Harpo Barrett <hbar0009@student.monash.edu>
Co-authored-by: Katherine Moffat <kmof0001@student.monash.edu>
Co-authored-by: Damien Jap <dami0004@student.monash.edu>