Fix: compile css class warnings (#13)
When running npx tailwindcss -i src/app.css we were getting two warnings.

After removing text-primary in the <a> in our app.css, the css classes are now compiling. (We're gonna use <a> for routing, so better to have a default style. But client still wants hyperlinks to be purple and underlined

However we are stilling getting and warning saying that our content section in tailwind.config.cjs is empty. But it is not empty, see line 2 tailwind.config.cjs

As far as I know even if we define all the places Tailwind classes are used, it still shows this warning according to this: https://stackoverflow.com/questions/70319979/purpose-of-content-in-tailwind-config

Otherwise following the docs https://tailwindcss.com/docs/content-configuration, I've also included daisyUI in our content, which should cover all places where Tailwind classes are used.

Co-authored-by: Harpo Barrett <61974761+hbar0009@users.noreply.github.com>